+++ src/App.js
@@ -1,28 +1,47 @@
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import "./App.css";
 import Header from "./components/Header/Header";
-import SimpleBottomNavigation from "./components/MainNav";
+import MainNav from "./components/MainNav";
 import Trending from "./components/Pages/Trending/Trending";
 import Movies from "./components/Pages/Movies/Movies";
 import Search from "./components/Pages/Search/Search";
 import Series from "./components/Pages/Series/Series";
+import { ThemeProvider, createMuiTheme } from "@material-ui/core";
 
 function App() {
+  const darkTheme = createMuiTheme({
+    palette: {
+      type: "dark",
+      primary: {
+        main: "#fff",
+      },
+    },
+  });
   return (
     <Router>
-      <Header />
-      <div className="App">
-        <container>
-          <Switch>
-            <Route path="/" component={Trending} exact />
-            <Route path="/movies" component={Movies} />
-            <Route path="/series" component={Series} />
-            <Route path="/search" component={Search} />
-          </Switch>
-        </container>
-      </div>
+      <ThemeProvider theme={darkTheme}>
+        <Header />
+        <div className="App">
+          <container>
+            <Switch>
+              <Route exact path="/">
+                <Trending />
+              </Route>
+              <Route path="/movies">
+                <Movies />
+              </Route>
+              <Route path="/series">
+                <Series />
+              </Route>
+              <Route path="/search">
+                <Search />
+              </Route>
+            </Switch>
+          </container>
+        </div>
 
-      <SimpleBottomNavigation />
+        <MainNav />
+      </ThemeProvider>
     </Router>
   );
 }
